Moves helper functions into Helper module
